Updated videos tracker
